+++ Portal/ServiceContainerCompilerPass/SetConfigurationAsParameterCompilerPass.php
@@ -9,7 +9,7 @@ use Heptacom\HeptaConnect\Core\Portal\PortalStackServiceContainerBuilder;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
-final class AddConfigurationAsParameterCompilerPass implements CompilerPassInterface
+final class SetConfigurationAsParameterCompilerPass implements CompilerPassInterface
 {
     public function __construct(
         private ?array $configuration
@@ -20,6 +20,12 @@ final class AddConfigurationAsParameterCompilerPass implements CompilerPassInter
     {
         $config = new PortalConfiguration($this->configuration ?? []);
         $keys = $config->keys();
+
+        foreach (\array_keys($container->getParameterBag()->all()) as $parameterName) {
+            if (\str_starts_with($parameterName, PortalStackServiceContainerBuilder::PORTAL_CONFIGURATION_PARAMETER_PREFIX)) {
+                $container->getParameterBag()->remove($parameterName);
+            }
+        }
 
         foreach ($keys as $key) {
             $container->setParameter($this->getParameterKey($key), $config->get($key));